refactor: controllers know only of domain

+++ foundation/gateway/internal/controller/regional/storage/get_sku_test.go
@@ -74,7 +74,7 @@ func TestStorageController_GetSKU(t *testing.T) {
 		require.NotNil(t, sku)
 		require.Equal(t, skuID, sku.Metadata.Name)
 		require.NotNil(t, sku.Spec)
-		require.Equal(t, 7500, sku.Spec.Iops)
+		require.Equal(t, int64(7500), sku.Spec.Iops)
 	})
 
 	t.Run("get_nonexistent", func(t *testing.T) {
